# specs/openapi.yaml
openapi: 3.0.0
info:
  title: Blog exercise
  version: 1.0.0
servers:
  - url: localhost:${HTTP_PORT}
paths:
  /posts:
    get:
      tags:
        - posts
      operationId: listAllPosts
      responses:
        '200':
          description: Get list of all posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseGetPost'

    post:
      tags:
        - posts
      operationId: createOnePost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
        required: true
      responses:
        '201':
          description: Create one post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

  /posts/{id}:
    put:
      tags:
        - posts
      operationId: updateOnePost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Update one post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          description: Invalid type or parameter send in body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - posts
      operationId: deleteOnePost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delete one post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          description: Invalid type or parameter send in body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /comments:
    get:
      tags:
        - comments
      operationId: listAllCommentsForAPost
      parameters:
        - name: postId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Get list of all comments for a post
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseGetComment'

    post:
      tags:
        - comments
      operationId: createOneComment
      parameters:
        - name: postId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
        required: true
      responses:
        '201':
          description: Create one comment for a post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

  /comments/{id}:
    put:
      tags:
        - comments
      operationId: updateOneComment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
        required: true
      responses:
        '200':
          description: Update one comment for a post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          description: Invalid type or parameter send in body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - comments
      operationId: deleteOneComment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delete one comment for a post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          description: Invalid type or parameter send in body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Response:
      type: object
      properties:
        response:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string

    CreatePostRequest:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        post:
          type: string
        likes:
          type: number

    ResponseGetPost:
      type: object
      properties:
        title:
          type: string
        post:
          type: string
        likes:
          type: number

    CreateCommentRequest:
      type: object
      properties:
        id:
          type: number
        comment:
          type: string

    UpdateCommentRequest:
      type: object
      properties:
        comment:
          type: string

    ResponseGetComment:
      type: object
      properties:
        comment:
          type: string